requirements.txt
streamlit
langchain
sentence-transformers
chromadb
beautifulsoup4
bs4
groq
python-dotenv